+++ Application/Business/Forecasts/IForecastValue.cs
@@ -6,14 +6,14 @@ namespace Application.Business.Forecasts
     public interface IForecastValue
     {
         DateTime DateTime { get; set; }
-        decimal Forecast { get; set; }
+        double Forecast { get; set; }
         double AskingPrice { get; set; }
         double BiddingPrice { get; set; }
-        decimal InstrumentBlock { get; set; }
-        decimal ShortForecast { get; }
-        decimal LongForecast { get; }
-        decimal MediumForecast { get; }
+        double InstrumentBlock { get; set; }
+        double ShortForecast { get; }
+        double LongForecast { get; }
+        double MediumForecast { get; }
         //Instrument Instrument { get; set; }
-        decimal CalculateForecast();
+        double CalculateForecast();
     }
 }
